+++ Lidgren.Network/NetUnreliableSenderChannel.cs
@@ -65,7 +65,7 @@ namespace Lidgren.Network
 		}
 
 		// call this regularely
-		internal override void SendQueuedMessages(double now)
+		internal override void SendQueuedMessages(float now)
 		{
 			int num = GetAllowedSends();
 			if (num < 1)
@@ -98,7 +98,7 @@ namespace Lidgren.Network
 		
 		// remoteWindowStart is remote expected sequence number; everything below this has arrived properly
 		// seqNr is the actual nr received
-		internal override void ReceiveAcknowledge(double now, int seqNr)
+		internal override void ReceiveAcknowledge(float now, int seqNr)
 		{
 			if (m_doFlowControl == false)
 			{
